Add files via upload
Added a conditional statement which checks if the savefile "phq9.csv" is present in the local directory, and if not, creates a new "phq9.csv" with headers. 

Also separated the date and the time, so that analysis over different times could be made easier, as well as analysis over days/months etc.